+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 ### Fix
 
-- 🔊 Add an option to delete uploaded custom Adhan audio files [#103](https://github.com/AbdelrahmanBayoumi/Azkar-App/issues/103)
+- 🔊 Add a delete button on each uploaded custom Adhan audio in the Adhan list [#103](https://github.com/AbdelrahmanBayoumi/Azkar-App/issues/103)
 
 ## [1.3.0](https://github.com/AbdelrahmanBayoumi/Azkar-App/releases/tag/1.3.0) - 2025-05-06